Write useUserCommunities() hook.
